+++ docker/musl.sh
@@ -48,7 +48,7 @@ main() {
 
     hide_output make install "-j$(nproc)" \
         GCC_VER=9.2.0 \
-        MUSL_VER=1.2.0 \
+        MUSL_VER=1.1.24 \
         BINUTILS_VER=2.33.1 \
         DL_CMD='curl --retry 3 -sSfL -C - -o' \
         LINUX_HEADERS_SITE=$linux_headers_site \